split view in multiple components
