+++ src/components/DarkModeComponent.vue
